i3: use gnome-terminal for main terminal app
urxvt, while fabulous, has issues with glyphs. Or the glyphs have issues
with urxvt. In any event, this is becoming increasingly annoying and in
some cases problematic. So for now switching to gnome-terminal for dev
work. Leaving other terminal instances in i3 to use urxvt, because they
can actually be named..