mutt: initial script to support delayed mail
The use case is writing email with no internet connection, and with a
little bit of metadata in a colon-separated file we can send those
messages later when needed.

Later parts should include a systemd service/timer setup, and probably a
bit attention to edge cases in the script itself. But this is 'working'