git-url: fix relative_path
When using `##` it was matching too greedily. We had the same project
name `hyrax` in the relative path `charts/hyrax/...` and so it was
clipping all the way to there, thus creating a url that didn't resolve.

Switching to a single `#` should just match the first instance (the
project name itself) and remove that as the "prefix", and keeping the
rest.

Also, add documentatino for Parameter Expansion because I always forget
what it's called.